gerar_imagem_com_mapa/requirements.txt
numpy==2.2.5
pillow==11.2.1
